requirements.txt
pytesseract
pillow
face_recognition
